MODE-1706 Moved Connector SPI classes to separate package, and added JavaDoc.
