+++ modeshape-jcr/src/main/java/org/modeshape/jcr/JcrI18n.java
@@ -105,6 +105,7 @@ public final class JcrI18n {
     public static I18n couldNotStoreProperties;
     public static I18n couldNotStoreProperty;
     public static I18n couldNotGetMimeType;
+    public static I18n connectorIsReadOnly;
 
     public static I18n rootNodeHasNoParent;
     public static I18n rootNodeIsNotProperty;
