+++ modeshape-jcr/src/main/java/org/modeshape/jcr/federation/spi/ExtraPropertiesStore.java
@@ -21,7 +21,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.modeshape.jcr.federation;
+package org.modeshape.jcr.federation.spi;
 
 import java.util.Collections;
 import java.util.Map;
