+++ modeshape-jcr/src/main/resources/org/modeshape/jcr/JcrI18n.properties
@@ -95,6 +95,7 @@ fileConnectorCannotStoreFileThatIsExcluded = The file system external source "{0
 couldNotStoreProperties = The external source "{0}" cannot store these extra properties on node "{1}": {2}
 couldNotStoreProperty = The external source "{0}" cannot store the extra property "{2}" on node "{1}"
 couldNotGetMimeType = The external source "{0}" cannot get the MIME type on node "{1}": {2}
+connectorIsReadOnly = The external source "{0}" is read-only and cannot update the node with identifier "{1}".
 
 rootNodeHasNoParent = The root node has no parent node
 rootNodeIsNotProperty = The root path "/" refers to the root node, not a property
