+++ modeshape-jcr/src/test/java/org/modeshape/connector/filesystem/InclusionExclusionFilenameFilterTest.java
@@ -27,7 +27,6 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
-import org.modeshape.connector.filesystem.FileSystemConnector.InclusionExclusionFilenameFilter;
 
 /**
  * @author johnament
